Some nits, cargo fmt

+++ emulator/cbinding/src/lib.rs
@@ -291,7 +291,7 @@ pub unsafe extern "C" fn emulator_init(
         } else {
             Some(config.i3c_port as u16)
         },
-        device_security_state: DeviceLifecycle::try_from(config.device_security_state as u32)
+        device_security_state: DeviceLifecycle::try_from(config.device_security_state)
             .unwrap_or(DeviceLifecycle::Production) as u32,
         vendor_pk_hash: convert_optional_c_string(config.vendor_pk_hash),
         vendor_pqc_type: caliptra_image_types::FwVerificationPqcKeyType::from_u8(
